Moved notification modals into a subfolder
