+++ assets/src/components/modal.tsx
@@ -3,10 +3,10 @@ import { SocketContext } from "../contexts/socketContext"
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
 import { ConnectionStatus } from "../hooks/useSocket"
 import DisconnectedModal from "./disconnectedModal"
-import InactiveNotificationModal from "./inactiveNotificationModal"
-import NotificationLoadingModal from "./notificationLoadingModal"
-import ChelseaRaisedNotificationModal from "./chelseaRaisedNotificationModal"
-import ChelseaLoweredNotificationModal from "./chelseaLoweredNotificationModal"
+import InactiveNotificationModal from "./notificationModals/inactiveNotificationModal"
+import NotificationLoadingModal from "./notificationModals/notificationLoadingModal"
+import ChelseaRaisedNotificationModal from "./notificationModals/chelseaRaisedNotificationModal"
+import ChelseaLoweredNotificationModal from "./notificationModals/chelseaLoweredNotificationModal"
 import CreatePresetModal from "./inputModals/createPresetModal"
 import SavePresetModal from "./inputModals/savePresetModal"
 import DeletePresetModal from "./inputModals/deletePresetModal"
