+++ assets/src/components/notificationModals/inactiveNotificationModal.tsx
@@ -1,13 +1,13 @@
 import React, { useContext } from "react"
-import Loading from "./loading"
-import { StateDispatchContext } from "../contexts/stateDispatchContext"
-import { useMinischeduleRuns } from "../hooks/useMinischedule"
-import { closeIcon } from "../helpers/icon"
-import { Activity, Run, Piece, Trip } from "../minischedule"
-import { Notification, RunId } from "../realtime.d"
-import { setNotification } from "../state"
-import { now, serviceDaySeconds } from "../util/dateTime"
-import { title } from "./notificationCard"
+import Loading from "../loading"
+import { StateDispatchContext } from "../../contexts/stateDispatchContext"
+import { useMinischeduleRuns } from "../../hooks/useMinischedule"
+import { closeIcon } from "../../helpers/icon"
+import { Activity, Run, Piece, Trip } from "../../minischedule"
+import { Notification, RunId } from "../../realtime.d"
+import { setNotification } from "../../state"
+import { now, serviceDaySeconds } from "../../util/dateTime"
+import { title } from "../notificationCard"
 
 type RunScheduleRelationship = "current" | "break" | "past"
